+++ resources/forms-ui/src/app/pages/dashboard/admin/admin-form-lists-page/admin-form-lists-page.component.ts
@@ -7,6 +7,7 @@ import { FormsService } from 'src/app/services/forms/forms.service';
 import { ListViewService } from 'src/app/services/view/list-view.service';
 import { LocalStorageService } from 'src/app/services/storage/local-storage.service';
 import { ClientService } from 'src/app/services/client/client.service';
+import { CompanyService } from 'src/app/services/company/company.service';
 
 @Component({
   selector: 'app-admin-form-lists-page',
@@ -37,6 +38,7 @@ export class AdminFormListsPageComponent implements OnInit {
     private modalService: NgbModal,
     private formService: FormsService,
     private clientService: ClientService,
+    private companyService: CompanyService,
     private listViewService: ListViewService,
     private localStorage: LocalStorageService
   ) {
@@ -210,9 +212,9 @@ export class AdminFormListsPageComponent implements OnInit {
     );
   }
 
-  searchByFormCode() {
+  searchByFormName() {
     this.loading = true;
-    this.clientService.findFormsByCode(this.query).then(
+    this.companyService.findFormsByName(this.query, this.merchant_id).then(
       forms => {
         if (forms.length == 0) {
           this.loading = false;
@@ -233,9 +235,10 @@ export class AdminFormListsPageComponent implements OnInit {
     );
   }
 
-  searchByFormName() {
+  searchFilteredFormsByName() {
     this.loading = true;
-    this.clientService.findFormsByName(this.query).then(
+    const status = this.filterState == 'active' ? 1 : 0;
+    this.companyService.findFormsByNameAndStatus(this.query, this.merchant_id, status).then(
       forms => {
         if (forms.length == 0) {
           this.loading = false;
@@ -260,31 +263,19 @@ export class AdminFormListsPageComponent implements OnInit {
     if (e.key == 'Enter') {
       const allForms = this.formsList;
       if (this.query.length != 0) {
-        // we need to know whether the user is searching by a form code
-        // or the user is searching by a form name.
-        // First, check if its a form code.
-        console.log(this.query);
-        this.hasError = false;
-        this.formsList = [];
-        if (/\d/.test(this.query)) {
-          if (this.query.length == 5) {
-            // search by form code, based on the input
-            // the user might be searching by a form code.
-            console.log('searching by form code');
-            this.searchByFormCode();
-          }
-          else {
-            // the input contains a number but is more than 5 characters
-            // in length, this might be a form name.
-            console.log('searching by form name');
-            this.searchByFormName();
-          }
+        if (this.filterState == 'all') {
+          console.log(this.query);
+          this.hasError = false;
+          this.formsList = [];
+          console.log('searching ...');
+          this.searchByFormName();
         }
         else {
-          // since all our form codes includes digits, and this
-          // users input doesnt include a digit, search by form name.
-          console.log('searching by form name last');
-          this.searchByFormName();
+          console.log(this.query);
+          this.hasError = false;
+          this.formsList = [];
+          console.log('searching on filter');
+          this.searchFilteredFormsByName();
         }
       }
       else {
@@ -292,7 +283,7 @@ export class AdminFormListsPageComponent implements OnInit {
           this.hasNoData = false;
           this.foundNoForm = false;
           this.formsList = [];
-          console.log('hererereererere');
+          this.filterState = 'all';
           this.getAllForms();
         }
       }