+++ resources/forms-ui/src/app/pages/dashboard/admin/admin-view-front-desks-page/admin-view-front-desks-page.component.ts
@@ -1,6 +1,6 @@
-import { Component, OnInit } from '@angular/core';
 import * as _ from 'lodash';
 import { Router } from '@angular/router';
+import { Component, OnInit } from '@angular/core';
 import { UserTypes } from 'src/app/enums/user-types.enum';
 import { LocalStorageService } from 'src/app/services/storage/local-storage.service';